feat: Add AppController and StudentsController
This commit adds the AppController and StudentsController classes to the repository. These controllers handle the logic for the homepage and student-related routes in the server. The AppController includes a getHomepage function that sends a 'Hello Holberton School!' response, while the StudentsController includes functions for retrieving all students and students by major from a database. These additions enhance the codebase by providing essential functionality for handling server requests related to the homepage and students, improving the overall functionality and flexibility of the code.